upgrading the selenium version for selenium grid 4 (#663)
